test appendable quantization
