+++ lib/segment/tests/integration/main.rs
@@ -12,7 +12,7 @@ mod gpu_hnsw_test;
 mod hnsw_discover_test;
 mod hnsw_incremental_build;
 mod hnsw_quantized_search_test;
-mod multivector_appendable_quantization_test;
+mod appendable_quantization_test;
 mod multivector_filtrable_hnsw_test;
 mod multivector_hnsw_test;
 mod multivector_quantization_test;
